+++ cms-api/src/main/java/com/condation/cms/api/Constants.java
@@ -62,6 +62,9 @@ public abstract class Constants {
 		public static final String TEMPLATE = "template";
 		
 		public static final String TYPE = "type";
+		
+		public static final String ALIASES = "aliases";
+		public static final String ALIASES_REDIRECT = "aliases_redirect";
 	}
 	
 	public static class Folders {