+++ cms-api/src/main/java/com/condation/cms/api/content/RedirectContentResponse.java
@@ -27,5 +27,5 @@ package com.condation.cms.api.content;
  *
  * @author t.marx
  */
-public interface ContentResponse {
+public record RedirectContentResponse(String location, int status) implements ContentResponse {
 };